requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
jinja2==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.0.335
langchain-community==0.0.13
chromadb==0.4.18
# neo4j==5.14.1 # 使用しないため削除
pycozo>=0.7.6 # CozoDB Python クライアント
markdown==3.5.1
pydantic==2.4.2
pytest==7.4.3
httpx==0.25.1
